tests/test_up.py
# test_up.py

import random
import unittest
import numpy as np
import torch
import os
from collections import defaultdict
from combat_sim.combat_env import CombatEnv
from agents.up import MADDPG, Actor, Critic, PrioritizedReplayBuffer, MixedNoise
import torch.nn as nn

class TestMADDPGUpgraded(unittest.TestCase):
    """测试升级后的MADDPG (up.py) 的完整功能和性能"""

    @classmethod
    def setUpClass(cls):
        """设置测试环境和基本参数"""
        print("\n开始升级后的MADDPG测试...")
        cls.env_config = {
            'num_red': 3,
            'num_blue': 2,
            'max_steps': 120,
            'field_size': 1000.0,
            'attack_range': 100.0,
            'min_speed': 10.0,
            'max_speed': 30.0,
            'max_turn_rate': np.pi / 6,
            'hit_probability': 0.8,
            'num_threads': 8
        }
        cls.env = CombatEnv(**cls.env_config)
        
        # 获取观察空间维度
        initial_obs = cls.env.reset()
        cls.obs_dim = len(list(initial_obs.values())[0])
        cls.act_dim = 3  # [转向，速度，开火]
        
        # 创建保存目录
        cls.save_dir = './test_results'
        os.makedirs(cls.save_dir, exist_ok=True)

    def setUp(self):
        """每个测试用例开始前的设置"""
        self.maddpg = MADDPG(
            n_red=self.env_config['num_red'],
            n_blue=self.env_config['num_blue'],
            obs_dim=self.obs_dim,
            act_dim=self.act_dim,
            n_step=3,  # 设置n_step为3以测试多步奖励机制
            gamma=0.99  # 设置gamma为0.99
        )

    def test_01_model_initialization(self):
        """测试模型初始化，包括层归一化和网络结构"""
        print("\n测试模型初始化...")
        
        # 测试Actor网络
        actor = Actor(self.obs_dim, self.act_dim)
        self.assertIsInstance(actor, nn.Module)
        # 检查是否包含LayerNorm层
        layers = [module for module in actor.modules()]
        layer_norms = [module for module in layers if isinstance(module, nn.LayerNorm)]
        self.assertTrue(len(layer_norms) >= 2, "Actor网络中应包含至少2个LayerNorm层")
        
        # 测试Critic网络
        critic = Critic(self.obs_dim, self.act_dim, self.env_config['num_red'] + self.env_config['num_blue'])
        self.assertIsInstance(critic, nn.Module)
        # 检查是否包含LayerNorm层
        layers = [module for module in critic.modules()]
        layer_norms = [module for module in layers if isinstance(module, nn.LayerNorm)]
        self.assertTrue(len(layer_norms) >= 2, "Critic网络中应包含至少2个LayerNorm层")
        
        # 测试网络输出维度
        dummy_obs = torch.randn(1, self.obs_dim)
        actor_out = actor(dummy_obs)
        self.assertEqual(actor_out.shape, (1, self.act_dim))
        
        # 测试优先级经验回放缓冲区
        buffer = PrioritizedReplayBuffer(n_step=3, gamma=0.99, n_agents=self.env_config['num_red'] + self.env_config['num_blue'], 
                                         n_red=self.env_config['num_red'], n_blue=self.env_config['num_blue'],
                                         obs_dim=self.obs_dim, act_dim=self.act_dim)  # 初始化时传入n_step和gamma
        self.assertEqual(len(buffer), 0)
        
        print("模型初始化测试完成")

    def test_02_action_selection(self):
        """测试动作选择功能，包括Mixed噪声的添加"""
        print("\n测试动作选择...")
        
        obs = self.env.reset()
        
        # 测试动作选择
        actions = self.maddpg.select_actions(obs)
        
        # 验证动作格式和范围
        for agent_id, action in actions.items():
            self.assertEqual(len(action), self.act_dim)
            self.assertTrue(np.all(action >= -1))
            self.assertTrue(np.all(action <= 1))
        
        # 测试动作选择的噪声功能
        actions_with_noise = self.maddpg.select_actions(obs, noise_scale=0.1)
        actions_without_noise = self.maddpg.select_actions(obs, noise_scale=0.0)
        
        # 确保噪声产生了不同的动作
        for agent_id in actions_with_noise:
            if agent_id in actions_without_noise:
                self.assertFalse(np.allclose(actions_with_noise[agent_id], actions_without_noise[agent_id]),
                                 f"Agent {agent_id} 的动作在有噪声和无噪声时应不同")
        
        print("动作选择测试完成")

    def test_03_experience_storage(self):
        """测试经验存储功能，包括优先级的存储和多步奖励"""
        print("\n测试经验存储...")
        
        # 定义固定的奖励序列
        fixed_rewards = [
            {f'red_{i}': 1.0 for i in range(self.maddpg.n_red)} |
            {f'blue_{i}': 1.5 for i in range(self.maddpg.n_blue)},
            {f'red_{i}': 2.0 for i in range(self.maddpg.n_red)} |
            {f'blue_{i}': 2.5 for i in range(self.maddpg.n_blue)},
            {f'red_{i}': 3.0 for i in range(self.maddpg.n_red)} |
            {f'blue_{i}': 3.5 for i in range(self.maddpg.n_blue)}
        ]
        
        # 创建一个MockEnv类来模拟环境的step方法
        class MockEnv:
            def __init__(self, fixed_rewards, obs_dim):
                self.fixed_rewards = fixed_rewards
                self.current_step = 0
                self.obs_dim = obs_dim
            
            def reset(self):
                self.current_step = 0
                return {k: np.random.rand(self.obs_dim) for k in self.fixed_rewards[0].keys()}
            
            def step(self, actions):
                if self.current_step < len(self.fixed_rewards):
                    rewards = self.fixed_rewards[self.current_step]
                    done = self.current_step == len(self.fixed_rewards) - 1
                    self.current_step += 1
                    next_obs = {k: np.random.rand(self.obs_dim) for k in rewards.keys()}
                    return next_obs, rewards, done, {}
                else:
                    return {k: np.random.rand(self.obs_dim) for k in actions.keys()}, \
                        {k: 0.0 for k in actions.keys()}, True, {}
        
        # 创建mock_env实例
        mock_env = MockEnv(fixed_rewards, self.obs_dim)
        
        # 替换环境的step方法为mock_env.step
        original_step = self.env.step
        self.env.step = mock_env.step
        
        try:
            # Store transitions
            for rewards in fixed_rewards:
                obs = self.env.reset()
                actions = self.maddpg.select_actions(obs)
                next_obs, rewards, done, _ = self.env.step(actions)
                self.maddpg.store_transition(
                    obs,
                    self._flatten_dict(actions),
                    self._flatten_dict(rewards),
                    next_obs,
                    {k: float(done) for k in obs.keys()}
                )
                if done:
                    break
            
            # Now buffer should have at least one n-step transition
            self.assertGreater(len(self.maddpg.memory), 0, "积累了n步后，缓冲区应包含经验")
            first_experience = self.maddpg.memory.buffer[0]
            
            # Calculate expected rewards per agent
            expected_rewards = []
            for agent_id in self.maddpg.memory.agent_ids:
                expected_reward = sum([self.maddpg.gamma**j * fixed_rewards[j].get(agent_id, 0.0) 
                                    for j in range(len(fixed_rewards))])
                expected_rewards.append(expected_reward)
            
            # Verify rewards
            actual_rewards = first_experience.rews
            self.assertEqual(len(actual_rewards), self.maddpg.n_agents, "经验中的奖励数量应与智能体数量一致")
            for exp, act in zip(expected_rewards, actual_rewards):
                self.assertAlmostEqual(exp, act, places=5, msg="n步奖励计算错误")
        
        finally:
            # 恢复原始环境的step方法
            self.env.step = original_step
        
        print("经验存储测试完成")


    def _get_reward_at_step(self, step):
        """辅助方法，用于获取特定步的奖励"""
        # 这是一个简化的示例，实际实现应根据环境的奖励结构
        # 在此示例中，我们假设奖励在环境内部被记录
        # 您可能需要修改此方法以适应您的环境
        # 假设有5个智能体，分别为red_0, red_1, red_2, blue_0, blue_1
        # 返回固定奖励，例如：[1.0, 2.0, 3.0, 4.0, 5.0]
        return [1.0, 2.0, 3.0, 4.0, 5.0]  # 示例奖励列表，长度应等于n_agents

    def test_04_prioritized_sampling(self):
        """测试优先级经验回放的采样功能"""
        print("\n测试优先级经验回放的采样功能...")
        
        # 填充缓冲区
        for _ in range(10):
            obs = self.env.reset()
            actions = self.maddpg.select_actions(obs)
            next_obs, rewards, done, _ = self.env.step(actions)
            self.maddpg.store_transition(
                obs,
                self._flatten_dict(actions),
                self._flatten_dict(rewards),
                next_obs,
                {k: float(done) for k in obs.keys()}
            )
            # Accumulate n_step=3
            for _ in range(2):
                if done:
                    break
                actions = self.maddpg.select_actions(next_obs)
                next_obs, rewards, done, _ = self.env.step(actions)
                self.maddpg.store_transition(
                    next_obs,
                    self._flatten_dict(actions),
                    self._flatten_dict(rewards),
                    next_obs,
                    {k: float(done) for k in obs.keys()}
                )
        
        # 测试采样
        batch_size = 5
        try:
            obs_batch, act_batch, rew_batch, next_obs_batch, dones_batch, indices, weights = self.maddpg.memory.sample(batch_size)
            self.assertEqual(obs_batch.shape[0], batch_size)
            self.assertEqual(act_batch.shape[0], batch_size)
            self.assertEqual(rew_batch.shape[0], batch_size)
            self.assertEqual(next_obs_batch.shape[0], batch_size)
            self.assertEqual(dones_batch.shape[0], batch_size)
            self.assertEqual(weights.shape[0], batch_size)
        except TypeError as e:
            self.fail(f"PrioritizedReplayBuffer.sample 失败: {e}")
        
        print("优先级经验回放的采样功能测试完成")

    def test_05_training_step(self):
        """测试训练步骤，包括梯度裁剪和优先级更新"""
        print("\n测试训练步骤...")
        
        # 填充一些经验
        for _ in range(self.maddpg.batch_size + 10):
            obs = self.env.reset()
            actions = self.maddpg.select_actions(obs)
            next_obs, rewards, done, _ = self.env.step(actions)
            self.maddpg.store_transition(
                obs,
                self._flatten_dict(actions),
                self._flatten_dict(rewards),
                next_obs,
                {k: float(done) for k in obs.keys()}
            )
            # Accumulate n_step=3
            for _ in range(2):
                if done:
                    break
                actions = self.maddpg.select_actions(next_obs)
                next_obs, rewards, done, _ = self.env.step(actions)
                self.maddpg.store_transition(
                    next_obs,
                    self._flatten_dict(actions),
                    self._flatten_dict(rewards),
                    next_obs,
                {k: float(done) for k in obs.keys()}
                )
        
        # 记录初始参数进行对比
        print("\n记录初始网络参数...")
        initial_critic_params = [param.clone() for param in self.maddpg.agents[0].critic.parameters()]
        initial_actor_params = [param.clone() for param in self.maddpg.agents[0].actor.parameters()]
        
        print("\n执行训练步骤...")
        self.maddpg.train()
        
        print("\n检查参数更新情况...")
        # 详细检查参数更新
        for i, (initial, updated) in enumerate(zip(initial_critic_params, self.maddpg.agents[0].critic.parameters())):
            param_changed = not torch.equal(initial, updated)
            print(f"Critic 参数组 {i}: {'已更新' if param_changed else '未更新'}")
            self.assertFalse(torch.equal(initial, updated), f"Critic参数组 {i} 应在训练后更新")
            if param_changed:
                diff_norm = torch.norm(initial - updated).item()
                print(f"  参数变化范数: {diff_norm:.6f}")
        
        for i, (initial, updated) in enumerate(zip(initial_actor_params, self.maddpg.agents[0].actor.parameters())):
            param_changed = not torch.equal(initial, updated)
            print(f"Actor 参数组 {i}: {'已更新' if param_changed else '未更新'}")
            self.assertFalse(torch.equal(initial, updated), f"Actor参数组 {i} 应在训练后更新")
            if param_changed:
                diff_norm = torch.norm(initial - updated).item()
                print(f"  参数变化范数: {diff_norm:.6f}")
        
        print("\n检查梯度裁剪情况...")
        # 检查梯度裁剪并输出详细信息
        for agent_idx, agent in enumerate(self.maddpg.agents):
            print(f"\n=== Agent {agent_idx} 的梯度信息 ===")
            
            print("\nCritic网络梯度:")
            for name, param in agent.critic.named_parameters():
                if param.grad is not None:
                    grad_norm = torch.norm(param.grad).item()
                    print(f"层: {name:20} - 梯度范数: {grad_norm:.4f}")
                    try:
                        self.assertTrue(grad_norm <= 3.0 + 1e-4,  # 根据up.py中的梯度裁剪范围调整
                                        f"Critic梯度超出限制: {grad_norm:.4f} > 3.0")
                    except AssertionError as e:
                        print(f"警告: {str(e)}")
            
            print("\nActor网络梯度:")
            for name, param in agent.actor.named_parameters():
                if param.grad is not None:
                    grad_norm = torch.norm(param.grad).item()
                    print(f"层: {name:20} - 梯度范数: {grad_norm:.4f}")
                    try:
                        self.assertTrue(grad_norm <= 1.5 + 1e-4,  # 根据up.py中的梯度裁剪范围调整
                                        f"Actor梯度超出限制: {grad_norm:.4f} > 1.5")
                    except AssertionError as e:
                        print(f"警告: {str(e)}")
        
        print("\n训练步骤测试完成")

    def test_06_model_save_load(self):
        """测试模型保存和加载，包括目标网络"""
        print("\n测试模型保存和加载...")
        
        save_path = os.path.join(self.save_dir, 'test_model_upgraded.pt')
        
        # 保存模型
        self.maddpg.save(save_path)
        self.assertTrue(os.path.exists(save_path), "模型保存路径应存在")
        
        # 记录当前Actor参数
        original_actor_params = [agent.actor.state_dict() for agent in self.maddpg.agents]
        
        # 修改模型参数
        for agent in self.maddpg.agents:
            for param in agent.actor.parameters():
                param.data += torch.randn_like(param)
        
        # 加载模型
        self.maddpg.load(save_path)
        
        # 验证加载后的参数与保存时一致
        loaded_actor_params = [agent.actor.state_dict() for agent in self.maddpg.agents]
        for original, loaded in zip(original_actor_params, loaded_actor_params):
            for key in original:
                np.testing.assert_array_almost_equal(original[key].numpy(), loaded[key].numpy(),
                                                     err_msg="加载后的Actor参数应与保存时一致")
        
        print("模型保存和加载测试完成")

    def test_07_mixed_noise_behavior(self):
        """测试MixedNoise类的行为"""
        print("\n测试MixedNoise类的行为...")
        
        noise = MixedNoise(size=self.act_dim, alpha=0.5)
        noise.reset()
        
        # 采样多个噪声
        samples = []
        # 增加暖机步骤以稳定OU噪声
        for _ in range(1000):
            noise.sample()
        for _ in range(1000):
            samples.append(noise.sample())
        samples = np.array(samples)
        
        # 检查噪声的均值接近mu
        mu = noise.mu
        mean_noise = samples.mean(axis=0)
        self.assertTrue(np.allclose(mean_noise, mu, atol=0.1), "MixedNoise的均值应接近mu")
        
        # 检查噪声的方差合理
        expected_var = (noise.alpha * (noise.sigma**2) / (2 * noise.theta)) + \
                    (1 - noise.alpha) * (noise.sigma_gaussian**2)
        expected_var = np.array([expected_var] * self.act_dim)
        var_noise = samples.var(axis=0)
        # 允许一定的容差
        for var, exp_var in zip(var_noise, expected_var):
            self.assertAlmostEqual(var, exp_var, delta=0.2 * exp_var, msg="MixedNoise方差不符合预期")
        
        print("MixedNoise类的行为测试完成")

    def test_08_multistep_rewards(self):
        """测试多步奖励机制的正确性"""
        print("\n测试多步奖励机制的正确性...")
        
        # Reset environment and create a specific sequence of rewards
        self.env.reset()
        fixed_rewards = [
            {f'red_{i}': 1.0 for i in range(self.env_config['num_red'])} |
            {f'blue_{i}': 1.5 for i in range(self.env_config['num_blue'])},
            {f'red_{i}': 2.0 for i in range(self.env_config['num_red'])} |
            {f'blue_{i}': 2.5 for i in range(self.env_config['num_blue'])},
            {f'red_{i}': 3.0 for i in range(self.env_config['num_red'])} |
            {f'blue_{i}': 3.5 for i in range(self.env_config['num_blue'])}
        ]
        
        # Override env.step to return fixed rewards
        original_step = self.env.step
        class MockEnv:
            def __init__(self, fixed_rewards):
                self.fixed_rewards = fixed_rewards
                self.current_step = 0
                self.n_agents = len(fixed_rewards[0])
            
            def reset(self):
                self.current_step = 0
                return {k: np.random.rand(self.obs_dim) for k in self.fixed_rewards[0].keys()}
            
            def step(self, actions):
                if self.current_step < len(self.fixed_rewards):
                    rewards = self.fixed_rewards[self.current_step]
                    done = self.current_step == len(self.fixed_rewards) - 1
                    self.current_step += 1
                    next_obs = {k: np.random.rand(self.obs_dim) for k in rewards.keys()}
                    return next_obs, rewards, done, {}
                else:
                    return {k: np.random.rand(self.obs_dim) for k in actions.keys()}, \
                           {k: 0.0 for k in actions.keys()}, True, {}
        
        mock_env = MockEnv(fixed_rewards)
        mock_env.obs_dim = self.obs_dim
        self.env.step = mock_env.step
        
        # Store transitions
        for r in fixed_rewards:
            obs = self.env.reset()
            actions = self.maddpg.select_actions(obs)
            next_obs, rewards, done, _ = self.env.step(actions)
            self.maddpg.store_transition(
                obs,
                self._flatten_dict(actions),
                self._flatten_dict(rewards),
                next_obs,
                {k: float(done) for k in obs.keys()}
            )
            if done:
                break
        
        # Now buffer should have at least one experience in multi-step
        self.assertGreater(len(self.maddpg.memory), 0, "缓冲区应包含n步经验")
        experience = self.maddpg.memory.buffer[0]
        
        # Calculate expected rewards per agent
        expected_rewards = []
        for agent_idx, agent_id in enumerate(self.maddpg.memory.agent_ids):
            expected_reward = sum([self.maddpg.gamma**i * fixed_rewards[i].get(agent_id, 0.0) 
                                   for i in range(len(fixed_rewards))])
            expected_rewards.append(expected_reward)
        
        # Verify rewards
        actual_rewards = experience.rews
        self.assertEqual(len(actual_rewards), self.maddpg.n_agents, "经验中的奖励数量应与智能体数量一致")
        for exp, act in zip(expected_rewards, actual_rewards):
            self.assertAlmostEqual(exp, act, places=5, msg="多步奖励计算错误")
        
        # Restore the original environment step
        self.env.step = original_step
        
        print("多步奖励机制的正确性测试完成")

    def test_09_buffer_dimension_consistency(self):
        """测试经验回放缓冲区的维度一致性"""
        print("\n测试经验回放缓冲区的维度一致性...")
        
        # 创建测试数据，确保包含所有智能体
        test_agents = {
            'red_0': np.random.rand(self.obs_dim),
            'red_1': np.random.rand(self.obs_dim),
            'red_2': np.random.rand(self.obs_dim),
            'blue_0': np.random.rand(self.obs_dim),
            'blue_1': np.random.rand(self.obs_dim)
        }
        obs = test_agents
        actions = {k: np.random.rand(self.act_dim) for k in test_agents.keys()}
        rewards = {k: random.random() for k in test_agents.keys()}
        next_obs = {k: np.random.rand(self.obs_dim) for k in test_agents.keys()}
        dones = {k: False for k in test_agents.keys()}
        
        # Store transitions to accumulate n_step=3
        self.maddpg.store_transition(obs, actions, rewards, next_obs, dones)
        for _ in range(2):
            obs = next_obs
            actions = {k: np.random.rand(self.act_dim) for k in test_agents.keys()}
            rewards = {k: random.random() for k in test_agents.keys()}
            next_obs = {k: np.random.rand(self.obs_dim) for k in test_agents.keys()}
            dones = {k: False for k in test_agents.keys()}
            self.maddpg.store_transition(obs, actions, rewards, next_obs, dones)
        
        # 检查采样的维度
        try:
            obs_batch, act_batch, rew_batch, next_obs_batch, done_batch, indices, weights = \
                self.maddpg.memory.sample(1)
            
            # 验证维度
            n_agents = self.maddpg.n_agents  # 使用MADDPG实例的智能体数量
            
            # 检查 done_batch 的维度
            done_reshaped = done_batch.view(1, n_agents)
            self.assertEqual(done_reshaped.shape, (1, n_agents),
                            f"Done batch shape mismatch: expected (1, {n_agents}), got {done_reshaped.shape}")
            
            print("经验回放缓冲区维度一致性测试通过")
            
        except Exception as e:
            self.fail(f"维度一致性测试失败: {str(e)}")

    def test_10_mixed_noise_behavior(self):
        """测试MixedNoise是否正确组合OU噪声和高斯噪声"""
        print("\n测试MixedNoise是否正确组合OU噪声和高斯噪声...")
        
        size = self.act_dim
        noise = MixedNoise(size=size, alpha=0.5)
        noise.reset()
        
        # Sample a number of noise samples
        samples = []
        # 增加暖机步骤以稳定OU噪声
        for _ in range(1000):
            noise.sample()
        for _ in range(1000):
            samples.append(noise.sample())
        samples = np.array(samples)
        
        # Compute statistics
        # 由于MixedNoise的组合方式复杂，这里主要验证噪声的均值和方差
        sample_mean = samples.mean(axis=0)
        sample_var = samples.var(axis=0)
        
        # Check that mean is approximately mu
        self.assertTrue(np.allclose(sample_mean, noise.mu, atol=0.1), "MixedNoise的均值应接近mu")
        
        # Expected variance calculation
        expected_var = (noise.alpha * (noise.sigma**2) / (2 * noise.theta)) + \
                    (1 - noise.alpha) * (noise.sigma_gaussian**2)
        expected_var = np.array([expected_var] * size)
        
        # Allow some tolerance
        for var, exp_var in zip(sample_var, expected_var):
            self.assertAlmostEqual(var, exp_var, delta=0.2 * exp_var, msg="MixedNoise方差不符合预期")
        
        print("MixedNoise的组合行为测试完成")


    def test_11_multistep_rewards(self):
        """测试多步奖励机制是否正确应用到训练"""
        print("\n测试多步奖励机制是否正确应用到训练...")
        
        # Initialize a new MADDPG with n_step=3
        maddpg = MADDPG(
            n_red=self.env_config['num_red'],
            n_blue=self.env_config['num_blue'],
            obs_dim=self.obs_dim,
            act_dim=self.act_dim,
            n_step=3,
            gamma=0.99
        )
        
        # Create a mock environment where we can control rewards
        class MockEnv:
            def __init__(self, fixed_rewards):
                self.fixed_rewards = fixed_rewards
                self.current_step = 0
                self.n_agents = len(fixed_rewards[0])
            
            def reset(self):
                self.current_step = 0
                return {k: np.random.rand(maddpg.obs_dim) for k in self.fixed_rewards[0].keys()}
            
            def step(self, actions):
                if self.current_step < len(self.fixed_rewards):
                    rewards = self.fixed_rewards[self.current_step]
                    done = self.current_step == len(self.fixed_rewards) - 1
                    self.current_step += 1
                    next_obs = {k: np.random.rand(maddpg.obs_dim) for k in rewards.keys()}
                    return next_obs, rewards, done, {}
                else:
                    return {k: np.random.rand(maddpg.obs_dim) for k in actions.keys()}, \
                           {k: 0.0 for k in actions.keys()}, True, {}
        
        fixed_rewards = [
            {f'red_{i}': 1.0 for i in range(maddpg.n_red)} |
            {f'blue_{i}': 1.5 for i in range(maddpg.n_blue)},
            {f'red_{i}': 2.0 for i in range(maddpg.n_red)} |
            {f'blue_{i}': 2.5 for i in range(maddpg.n_blue)},
            {f'red_{i}': 3.0 for i in range(maddpg.n_red)} |
            {f'blue_{i}': 3.5 for i in range(maddpg.n_blue)}
        ]
        mock_env = MockEnv(fixed_rewards)
        mock_env.obs_dim = self.obs_dim
        original_step = self.env.step
        self.env.step = mock_env.step
        
        # Store transitions
        obs = self.env.reset()
        for _ in range(3):
            actions = maddpg.select_actions(obs)
            next_obs, rewards, done, _ = self.env.step(actions)
            maddpg.store_transition(
                obs,
                self._flatten_dict(actions),
                self._flatten_dict(rewards),
                next_obs,
                {k: float(done) for k in obs.keys()}
            )
            obs = next_obs
            if done:
                break
        
        # Now, there should be one experience in the buffer with n_step reward
        self.assertEqual(len(maddpg.memory), 1, "缓冲区应包含1个n步经验")
        experience = maddpg.memory.buffer[0]
        
        # Calculate expected rewards per agent
        expected_rewards = []
        for agent_id in maddpg.memory.agent_ids:
            expected_reward = sum([maddpg.gamma**i * fixed_rewards[i].get(agent_id, 0.0) 
                                   for i in range(len(fixed_rewards))])
            expected_rewards.append(expected_reward)
        
        # Verify rewards
        actual_rewards = experience.rews
        self.assertEqual(len(actual_rewards), maddpg.n_agents, "经验中的奖励数量应与智能体数量一致")
        for exp, act in zip(expected_rewards, actual_rewards):
            self.assertAlmostEqual(exp, act, places=5, msg="多步奖励计算错误")
        
        # Restore the original environment step
        self.env.step = original_step
        
        print("多步奖励机制正确应用到训练测试完成")

    def _run_training_session(self, n_episodes):
        """运行训练会话并返回统计数据"""
        stats = {
            'episode_rewards': [],
            'episode_lengths': [],
            'red_win_rates': [],
            'red_survival_rates': [],
            'blue_survival_rates': []
        }
        
        for episode in range(n_episodes):
            obs = self.env.reset()
            episode_reward = 0
            steps = 0
            
            while True:
                actions = self.maddpg.select_actions(obs)
                next_obs, rewards, done, info = self.env.step(actions)
                
                self.maddpg.store_transition(
                    obs,
                    self._flatten_dict(actions),
                    self._flatten_dict(rewards),
                    next_obs,
                    {k: float(done) for k in obs.keys()}
                )
                
                self.maddpg.train()
                
                episode_reward += sum(rewards.values())
                steps += 1
                obs = next_obs
                
                if done:
                    break
            
            # 更新统计数据
            stats['episode_rewards'].append(episode_reward)
            stats['episode_lengths'].append(steps)
            
            # 计算红方胜率
            red_win = info['blue_alive'] == 0
            stats['red_win_rates'].append(float(red_win))
            
            # 计算存活率
            red_alive = info.get('red_alive', 0)
            blue_alive = info.get('blue_alive', 0)
            
            stats['red_survival_rates'].append(red_alive / self.env_config['num_red'])
            stats['blue_survival_rates'].append(blue_alive / self.env_config['num_blue'])
        
        return stats

    def _flatten_dict(self, d):
        """将字典展平为列表，按照红方后蓝方的顺序"""
        result = []
        for i in range(self.env_config['num_red']):
            key = f'red_{i}'
            if key in d:
                result.append(d[key])
            else:
                result.append([0.0] * self.act_dim)  # 填充零向量
        for i in range(self.env_config['num_blue']):
            key = f'blue_{i}'
            if key in d:
                result.append(d[key])
            else:
                result.append([0.0] * self.act_dim)  # 填充零向量
        return result

    @classmethod
    def tearDownClass(cls):
        """清理测试环境"""
        cls.env.close()
        print("\n升级后的MADDPG测试完成!")

def main():
    # 运行所有测试
    unittest.main(verbosity=2)

if __name__ == '__main__':
    main()
